requirements.txt
aniso8601==9.0.1
attrs==22.1.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
Flask==2.2.2
flask-restx==1.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
mysqlclient==2.1.1
pycparser==2.21
PyMySQL==1.0.2
pyrsistent==0.19.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
